+++ crates/front/src/compile/elab.rs
@@ -1082,7 +1082,7 @@ pub(crate) fn elab_expr<'a>(
             let ind_name = infer_inductive(ctx, scope, scrutinee).ok_or_else(|| {
                 CompileError::elab(
                     ErrorKind::ElabMatchNotInductive,
-                    "`match` 的被匹配项不是本文件里用 inductive 声明的归纳类型（v1 只支持源内非递归枚举/结构体）",
+                    "`match` 的被匹配项不是已知的归纳类型（本文件用 inductive 声明，或 prelude 的 Nat）",
                     scrutinee.span(),
                 )
             })?;
@@ -1090,7 +1090,7 @@ pub(crate) fn elab_expr<'a>(
                 CompileError::elab(
                     ErrorKind::ElabMatchNotInductive,
                     format!(
-                        "`match` 的被匹配项类型 `{ind_name}` 不是本文件用 inductive 声明的类型"
+                        "`match` 的被匹配项类型 `{ind_name}` 不是已知的归纳类型（本文件用 inductive 声明，或 prelude 的 Nat）"
                     ),
                     scrutinee.span(),
                 )